+++ packages/x-date-pickers-pro/src/AdapterDateFnsJalaliV2/index.ts
@@ -1,0 +1,1 @@
+export { AdapterDateFnsJalali } from '@mui/x-date-pickers/AdapterDateFnsJalaliV2';
