--- packages/x-date-pickers-pro/src/AdapterDateFnsJalaliV3/index.ts
@@ -1,1 +1,0 @@
-export { AdapterDateFnsJalali } from '@mui/x-date-pickers/AdapterDateFnsJalaliV3';
