+++ packages/x-date-pickers-pro/src/AdapterDateFnsV2/index.ts
@@ -1,0 +1,1 @@
+export { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFnsV2';
