--- packages/x-date-pickers-pro/src/AdapterDateFnsV3/index.ts
@@ -1,1 +1,0 @@
-export { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFnsV3';
