+++ packages/x-date-pickers/src/AdapterDateFnsJalaliV2/index.ts
@@ -1,0 +1,1 @@
+export { AdapterDateFnsJalali } from './AdapterDateFnsJalaliV2';
