--- packages/x-date-pickers/src/AdapterDateFnsJalaliV3/index.ts
@@ -1,1 +1,0 @@
-export { AdapterDateFnsJalali } from './AdapterDateFnsJalaliV3';
