+++ packages/x-date-pickers/src/AdapterDateFnsV2/AdapterDateFnsV2.test.tsx
@@ -1,8 +1,8 @@
-import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFnsV3';
+import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFnsV2';
 import { describeGregorianAdapter } from 'test/utils/pickers/describeGregorianAdapter';
 import { fr } from 'date-fns/locale';
 
-describe('<AdapterDateFnsV3 />', () => {
+describe('<AdapterDateFnsV2 />', () => {
   describeGregorianAdapter(AdapterDateFns, {
     formatDateTime: 'yyyy-MM-dd HH:mm:ss',
     setDefaultTimezone: () => {},
