+++ packages/x-date-pickers/src/AdapterDateFnsV2/AdapterDateFnsV2.ts
@@ -1,55 +1,54 @@
 /* eslint-disable class-methods-use-this */
-// TODO remove when date-fns-v3 is the default
 // @ts-nocheck
-import { addDays } from 'date-fns/addDays';
-import { addSeconds } from 'date-fns/addSeconds';
-import { addMinutes } from 'date-fns/addMinutes';
-import { addHours } from 'date-fns/addHours';
-import { addWeeks } from 'date-fns/addWeeks';
-import { addMonths } from 'date-fns/addMonths';
-import { addYears } from 'date-fns/addYears';
-import { endOfDay } from 'date-fns/endOfDay';
-import { endOfWeek } from 'date-fns/endOfWeek';
-import { endOfYear } from 'date-fns/endOfYear';
-// @ts-ignore TODO remove when date-fns-v3 is the default
-import { format as dateFnsFormat, longFormatters } from 'date-fns/format';
-import { getDate } from 'date-fns/getDate';
-import { getDaysInMonth } from 'date-fns/getDaysInMonth';
-import { getHours } from 'date-fns/getHours';
-import { getMinutes } from 'date-fns/getMinutes';
-import { getMonth } from 'date-fns/getMonth';
-import { getSeconds } from 'date-fns/getSeconds';
-import { getMilliseconds } from 'date-fns/getMilliseconds';
-import { getWeek } from 'date-fns/getWeek';
-import { getYear } from 'date-fns/getYear';
-import { isAfter } from 'date-fns/isAfter';
-import { isBefore } from 'date-fns/isBefore';
-import { isEqual } from 'date-fns/isEqual';
-import { isSameDay } from 'date-fns/isSameDay';
-import { isSameYear } from 'date-fns/isSameYear';
-import { isSameMonth } from 'date-fns/isSameMonth';
-import { isSameHour } from 'date-fns/isSameHour';
-import { isValid } from 'date-fns/isValid';
-import { parse as dateFnsParse } from 'date-fns/parse';
-import { setDate } from 'date-fns/setDate';
-import { setHours } from 'date-fns/setHours';
-import { setMinutes } from 'date-fns/setMinutes';
-import { setMonth } from 'date-fns/setMonth';
-import { setSeconds } from 'date-fns/setSeconds';
-import { setMilliseconds } from 'date-fns/setMilliseconds';
-import { setYear } from 'date-fns/setYear';
-import { startOfDay } from 'date-fns/startOfDay';
-import { startOfMonth } from 'date-fns/startOfMonth';
-import { endOfMonth } from 'date-fns/endOfMonth';
-import { startOfWeek } from 'date-fns/startOfWeek';
-import { startOfYear } from 'date-fns/startOfYear';
-import { isWithinInterval } from 'date-fns/isWithinInterval';
-import { enUS } from 'date-fns/locale/en-US';
-// date-fns v2 does not export types
-// @ts-ignore TODO remove when date-fns-v3 is the default
-import { Locale as DateFnsLocale } from 'date-fns/locale/types';
+import addDays from 'date-fns/addDays';
+import addSeconds from 'date-fns/addSeconds';
+import addMinutes from 'date-fns/addMinutes';
+import addHours from 'date-fns/addHours';
+import addWeeks from 'date-fns/addWeeks';
+import addMonths from 'date-fns/addMonths';
+import addYears from 'date-fns/addYears';
+import endOfDay from 'date-fns/endOfDay';
+import endOfWeek from 'date-fns/endOfWeek';
+import endOfYear from 'date-fns/endOfYear';
+import dateFnsFormat from 'date-fns/format';
+import getDate from 'date-fns/getDate';
+import getDaysInMonth from 'date-fns/getDaysInMonth';
+import getHours from 'date-fns/getHours';
+import getMinutes from 'date-fns/getMinutes';
+import getMonth from 'date-fns/getMonth';
+import getSeconds from 'date-fns/getSeconds';
+import getMilliseconds from 'date-fns/getMilliseconds';
+import getWeek from 'date-fns/getWeek';
+import getYear from 'date-fns/getYear';
+import isAfter from 'date-fns/isAfter';
+import isBefore from 'date-fns/isBefore';
+import isEqual from 'date-fns/isEqual';
+import isSameDay from 'date-fns/isSameDay';
+import isSameYear from 'date-fns/isSameYear';
+import isSameMonth from 'date-fns/isSameMonth';
+import isSameHour from 'date-fns/isSameHour';
+import isValid from 'date-fns/isValid';
+import dateFnsParse from 'date-fns/parse';
+import setDate from 'date-fns/setDate';
+import setHours from 'date-fns/setHours';
+import setMinutes from 'date-fns/setMinutes';
+import setMonth from 'date-fns/setMonth';
+import setSeconds from 'date-fns/setSeconds';
+import setMilliseconds from 'date-fns/setMilliseconds';
+import setYear from 'date-fns/setYear';
+import startOfDay from 'date-fns/startOfDay';
+import startOfMonth from 'date-fns/startOfMonth';
+import endOfMonth from 'date-fns/endOfMonth';
+import startOfWeek from 'date-fns/startOfWeek';
+import startOfYear from 'date-fns/startOfYear';
+import isWithinInterval from 'date-fns/isWithinInterval';
+import defaultLocale from 'date-fns/locale/en-US';
+// @ts-ignore
+import longFormatters from 'date-fns/_lib/format/longFormatters';
 import { AdapterFormats, AdapterOptions, MuiPickersAdapter } from '../models';
 import { AdapterDateFnsBase } from '../AdapterDateFnsBase';
+
+type DateFnsLocale = typeof defaultLocale;
 
 declare module '@mui/x-date-pickers/models' {
   interface PickerValidDateLookup {
@@ -92,21 +91,15 @@ export class AdapterDateFns
       if (typeof addDays !== 'function') {
         throw new Error(
           [
-            `MUI: The \`date-fns\` package v2.x is not compatible with this adapter.`,
-            'Please, install v3.x or v4.x of the package or use the `AdapterDateFns` instead.',
+            'MUI: This adapter is only compatible with `date-fns` v2.x package versions.',
+            'Please, install v2.x of the package or use the `AdapterDateFns` instead.',
           ].join('\n'),
         );
       }
-      if (!longFormatters) {
-        throw new Error(
-          'MUI: The minimum supported `date-fns` package version compatible with this adapter is `3.2.x`.',
-        );
-      }
     }
-    super({ locale: locale ?? enUS, formats, longFormatters });
+    super({ locale: locale ?? defaultLocale, formats, longFormatters });
   }
 
-  // TODO: explicit return types can be removed once there is only one date-fns version supported
   public parse = (value: string, format: string): Date | null => {
     if (value === '') {
       return null;
