+++ packages/x-date-pickers/src/AdapterDateFnsV2/index.ts
@@ -1,0 +1,1 @@
+export { AdapterDateFns } from './AdapterDateFnsV2';
