--- packages/x-date-pickers/src/AdapterDateFnsV3/index.ts
@@ -1,1 +1,0 @@
-export { AdapterDateFns } from './AdapterDateFnsV3';
